requirements.txt
djangorestframework==3.4.7